Fix flaky async / thread pool tests
The previous `std::async()`-based approach to async calculations
assumed that as many threads as tasks would be spun up simultaneously.
This is one of the behaviours the new thread pool-based approach
intentionally changes.  These tests became flaky when run using the
number of available hardware threads on smaller machines (fewer than 6
threads).  This patch explicitly asks for as many workers as tasks in
these tests to restore the previous test semantics.